remove protocols from DocCDocumentation module
