+++ Sources/CMakeLists.txt
@@ -8,7 +8,6 @@ add_subdirectory(CCompletionScoring)
 add_subdirectory(CompletionScoring)
 add_subdirectory(Csourcekitd)
 add_subdirectory(Diagnose)
-add_subdirectory(DocCDocumentation)
 add_subdirectory(InProcessClient)
 add_subdirectory(LanguageServerProtocol)
 add_subdirectory(LanguageServerProtocolExtensions)
