+++ Sources/SourceKitLSP/Clang/ClangLanguageService.swift
@@ -11,7 +11,6 @@
 //===----------------------------------------------------------------------===//
 
 import BuildSystemIntegration
-import DocCDocumentation
 import Foundation
 import LanguageServerProtocol
 import LanguageServerProtocolExtensions
@@ -22,6 +21,10 @@ import SwiftExtensions
 import SwiftSyntax
 import TSCExtensions
 import ToolchainRegistry
+
+#if canImport(DocCDocumentation)
+import DocCDocumentation
+#endif
 
 #if os(Windows)
 import WinSDK
@@ -483,9 +486,11 @@ extension ClangLanguageService {
     return try await forwardRequestToClangd(req)
   }
 
+  #if canImport(DocCDocumentation)
   func doccDocumentation(_ req: DoccDocumentationRequest) async throws -> DoccDocumentationResponse {
     throw ResponseError.requestFailed(doccDocumentationError: .noDocumentation)
   }
+  #endif
 
   func symbolInfo(_ req: SymbolInfoRequest) async throws -> [SymbolDetails] {
     return try await forwardRequestToClangd(req)
