+++ Sources/SourceKitLSP/SourceKitLSPServer.swift
@@ -13,7 +13,6 @@
 import BuildServerProtocol
 import BuildSystemIntegration
 import Dispatch
-package import DocCDocumentation
 import Foundation
 import IndexStoreDB
 package import LanguageServerProtocol
@@ -28,6 +27,10 @@ package import ToolchainRegistry
 
 import struct TSCBasic.AbsolutePath
 import protocol TSCBasic.FileSystem
+
+#if canImport(DocCDocumentation)
+package import DocCDocumentation
+#endif
 
 /// Disambiguate LanguageServerProtocol.Language and IndexstoreDB.Language
 package typealias Language = LanguageServerProtocol.Language
@@ -76,7 +79,9 @@ package actor SourceKitLSPServer {
 
   package let documentManager = DocumentManager()
 
-  package let documentationManager = createDocCDocumentationManager()
+  #if canImport(DocCDocumentation)
+  package let documentationManager = DocCDocumentationManager()
+  #endif
 
   /// The `TaskScheduler` that schedules all background indexing tasks.
   ///
@@ -751,8 +756,10 @@ extension SourceKitLSPServer: QueueBasedMessageHandler {
       await self.handleRequest(for: request, requestHandler: self.declaration)
     case let request as RequestAndReply<DefinitionRequest>:
       await self.handleRequest(for: request, requestHandler: self.definition)
+    #if canImport(DocCDocumentation)
     case let request as RequestAndReply<DoccDocumentationRequest>:
       await self.handleRequest(for: request, requestHandler: self.doccDocumentation)
+    #endif
     case let request as RequestAndReply<DocumentColorRequest>:
       await self.handleRequest(for: request, requestHandler: self.documentColor)
     case let request as RequestAndReply<DocumentDiagnosticsRequest>:
@@ -1076,9 +1083,9 @@ extension SourceKitLSPServer {
       GetReferenceDocumentRequest.method: .dictionary(["version": .int(1)]),
       DidChangeActiveDocumentNotification.method: .dictionary(["version": .int(1)]),
     ]
-    if (await documentationManager.getRenderingSupport() != nil) {
-      experimentalCapabilities["textDocument/doccDocumentation"] = .dictionary(["version": .int(1)])
-    }
+    #if canImport(DocCDocumentation)
+    experimentalCapabilities["textDocument/doccDocumentation"] = .dictionary(["version": .int(1)])
+    #endif
 
     return ServerCapabilities(
       textDocumentSync: .options(
@@ -1469,7 +1476,9 @@ extension SourceKitLSPServer {
     // (e.g. Package.swift doesn't have build settings but affects build
     // settings). Inform the build system about all file changes.
     await workspaces.concurrentForEach { await $0.filesDidChange(notification.changes) }
-    await documentationManager.getRenderingSupport()?.filesDidChange(notification.changes)
+    #if canImport(DocCDocumentation)
+    await documentationManager.filesDidChange(notification.changes)
+    #endif
   }
 
   func setBackgroundIndexingPaused(_ request: SetOptionsRequest) async throws -> VoidResponse {
@@ -1561,6 +1570,7 @@ extension SourceKitLSPServer {
     return try await documentService(for: completionItemData.uri).completionItemResolve(request)
   }
 
+  #if canImport(DocCDocumentation)
   func doccDocumentation(
     _ req: DoccDocumentationRequest,
     workspace: Workspace,
@@ -1568,6 +1578,7 @@ extension SourceKitLSPServer {
   ) async throws -> DoccDocumentationResponse {
     return try await languageService.doccDocumentation(req)
   }
+  #endif
 
   func hover(
     _ req: HoverRequest,
